chore: (#25) data -> domain 직관적인 이름 사용
