+++ storties-search/src/main/java/org/example/stortiessearch/application/service/CreatePostService.java
@@ -6,8 +6,8 @@ import org.example.stortiessearch.application.event.CreatePostEvent;
 import org.example.stortiessearch.infrastructure.mq.producer.CreatePostProducer;
 import org.example.stortiessearch.global.authentication.AuthenticatedUserProvider;
 import org.example.stortiessearch.infrastructure.client.grpc.user.dto.AuthenticatedUser;
-import org.example.stortiessearch.data.persistence.post.CommandPostRepository;
-import org.example.stortiessearch.data.persistence.post.model.PostEntity;
+import org.example.stortiessearch.domain.post.CommandPostRepository;
+import org.example.stortiessearch.domain.post.model.PostEntity;
 import org.example.stortiessearch.application.service.dto.request.CreatePostRequest;
 import org.springframework.stereotype.Service;
