+++ storties-search/src/main/java/org/example/stortiessearch/application/service/DeletePostService.java
@@ -7,9 +7,9 @@ import org.example.stortiessearch.infrastructure.mq.producer.DeletePostProducer;
 import org.example.stortiessearch.global.authentication.AuthenticatedUserProvider;
 import org.example.stortiessearch.global.exception.error.ErrorCodes;
 import org.example.stortiessearch.infrastructure.client.grpc.user.dto.AuthenticatedUser;
-import org.example.stortiessearch.data.persistence.post.CommandPostRepository;
-import org.example.stortiessearch.data.persistence.post.QueryPostRepository;
-import org.example.stortiessearch.data.persistence.post.model.PostEntity;
+import org.example.stortiessearch.domain.post.CommandPostRepository;
+import org.example.stortiessearch.domain.post.QueryPostRepository;
+import org.example.stortiessearch.domain.post.model.PostEntity;
 import org.springframework.stereotype.Service;
 
 @Service
