+++ storties-search/src/main/java/org/example/stortiessearch/application/service/QueryPostsService.java
@@ -3,8 +3,8 @@ package org.example.stortiessearch.application.service;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.example.stortiessearch.application.service.dto.response.PostsResponse;
-import org.example.stortiessearch.data.persistence.post.QueryPostRepository;
-import org.example.stortiessearch.data.persistence.post.model.PostEntity;
+import org.example.stortiessearch.domain.post.QueryPostRepository;
+import org.example.stortiessearch.domain.post.model.PostEntity;
 import org.example.stortiessearch.infrastructure.cache.service.GetLikeCountService;
 import org.example.stortiessearch.infrastructure.cache.service.GetViewCountService;
 import org.springframework.stereotype.Service;
