+++ storties-search/src/main/java/org/example/stortiessearch/application/service/UnLikeService.java
@@ -2,7 +2,7 @@ package org.example.stortiessearch.application.service;
 
 import lombok.RequiredArgsConstructor;
 import org.example.stortiessearch.application.event.DecreasePostLikeEvent;
-import org.example.stortiessearch.data.persistence.post.QueryPostRepository;
+import org.example.stortiessearch.domain.post.QueryPostRepository;
 import org.example.stortiessearch.global.authentication.AuthenticatedUserProvider;
 import org.example.stortiessearch.infrastructure.mq.producer.DecreasePostLikeProducer;
 import org.springframework.stereotype.Component;
