+++ storties-search/src/main/java/org/example/stortiessearch/application/service/UpdatePostService.java
@@ -6,8 +6,8 @@ import lombok.RequiredArgsConstructor;
 import org.example.stortiessearch.application.event.UpdatePostEvent;
 import org.example.stortiessearch.application.service.dto.request.UpdatePostRequest;
 import org.example.stortiessearch.global.exception.error.ErrorCodes;
-import org.example.stortiessearch.data.persistence.post.model.PostEntity;
-import org.example.stortiessearch.data.persistence.post.repository.PostJpaRepository;
+import org.example.stortiessearch.domain.post.model.PostEntity;
+import org.example.stortiessearch.domain.post.repository.PostJpaRepository;
 import org.example.stortiessearch.infrastructure.client.grpc.user.dto.AuthenticatedUser;
 import org.example.stortiessearch.infrastructure.mq.producer.UpdatePostProducer;
 import org.example.stortiessearch.global.authentication.AuthenticatedUserProvider;
