+++ storties-search/src/main/java/org/example/stortiessearch/domain/post/model/PostEntity.java
@@ -1,4 +1,4 @@
-package org.example.stortiessearch.data.persistence.post.model;
+package org.example.stortiessearch.domain.post.model;
 
 import jakarta.persistence.CollectionTable;
 import jakarta.persistence.Column;
