+++ storties-search/src/main/java/org/example/stortiessearch/domain/post/model/PostViewEntity.java
@@ -1,4 +1,4 @@
-package org.example.stortiessearch.data.persistence.post.model;
+package org.example.stortiessearch.domain.post.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -11,7 +11,6 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -26,14 +25,12 @@ import org.hibernate.annotations.OnDeleteAction;
 @Builder
 @Getter
 @Table(
-    name = "tbl_post_like",
+    name = "tbl_post_view",
     indexes = {
-        @Index(name = "idx_like_log_post_id", columnList = "post_id")
-    },
-    uniqueConstraints = {
-        @UniqueConstraint(name = "uk_like_post_id_user_id", columnNames = {"post_id", "user_id"})
-    })
-public class PostLikeEntity {
+        @Index(name = "idx_view_post_id", columnList = "post_id")
+    }
+)
+public class PostViewEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
@@ -47,13 +44,13 @@ public class PostLikeEntity {
     @Column(name = "user_id")
     private Long userId;
 
-    @Column(name = "liked_at")
-    private LocalDateTime likedAt;
+    @Column(name = "viewed_at")
+    private LocalDateTime viewedAt;
 
     @PrePersist
     public void prePersist() {
-        if (likedAt == null) {
-            this.likedAt = LocalDateTime.now();
+        if (viewedAt == null) {
+            this.viewedAt = LocalDateTime.now();
         }
     }
 }
