+++ storties-search/src/main/java/org/example/stortiessearch/domain/post/repository/PostLikeJpaRepository.java
@@ -1,6 +1,6 @@
-package org.example.stortiessearch.data.persistence.post.repository;
+package org.example.stortiessearch.domain.post.repository;
 
-import org.example.stortiessearch.data.persistence.post.model.PostLikeEntity;
+import org.example.stortiessearch.domain.post.model.PostLikeEntity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface PostLikeJpaRepository extends CrudRepository<PostLikeEntity, Long> {
