+++ storties-search/src/main/java/org/example/stortiessearch/domain/post/repository/PostViewJpaRepository.java
@@ -1,6 +1,6 @@
-package org.example.stortiessearch.data.persistence.post.repository;
+package org.example.stortiessearch.domain.post.repository;
 
-import org.example.stortiessearch.data.persistence.post.model.PostViewEntity;
+import org.example.stortiessearch.domain.post.model.PostViewEntity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface PostViewJpaRepository extends CrudRepository<PostViewEntity, Long> {
